+++ ubo_app/engines/abstraction/background_running_mixin.py
@@ -11,41 +11,31 @@ from ubo_app.logger import logger
 from ubo_app.store.main import store
 from ubo_app.store.services.notifications import Notification, NotificationsAddAction
 from ubo_app.utils.async_ import create_task
-from ubo_app.utils.async_evicting_queue import AsyncEvictingQueue
 from ubo_app.utils.error_handlers import report_service_error
 
 if TYPE_CHECKING:
     import asyncio
 
-    from ubo_app.store.services.speech_recognition import (
-        SpeechRecognitionEngineName,
-    )
-
 
 class BackgroundRunningMixin(abc.ABC):
-    """Base class for recognition engines."""
+    """Base class for third-party background running engines."""
 
-    name: SpeechRecognitionEngineName
+    name: str
     label: str
-    input_chunks_queue: AsyncEvictingQueue[bytes]
 
     def __init__(
         self,
         *,
-        name: SpeechRecognitionEngineName,
+        name: str,
         label: str,
     ) -> None:
-        """Initialize the recognition engine."""
+        """Initialize `BackgroundRunningMixin`."""
         self.name = name
         self.label = label
-        self.input_chunks_queue = AsyncEvictingQueue(maxsize=5)
+        self._task = None
         self._run_lock = threading.Lock()
         self._is_running: bool = False
         super().__init__()
-
-    async def queue_audio_chunk(self, chunk: bytes) -> None:
-        """Queue a chunk of audio data for processing."""
-        await self.input_chunks_queue.put(chunk)
 
     @final
     def _task_done_callback(self, task: asyncio.Task[None]) -> None:
@@ -53,7 +43,7 @@ class BackgroundRunningMixin(abc.ABC):
         self._task = None
         if not task.cancelled() and task.exception():
             logger.exception(
-                'Speech recognition engine task failed',
+                'An error occurred while running the engine',
                 extra={
                     'engine_name': self.name,
                 },
@@ -63,11 +53,10 @@ class BackgroundRunningMixin(abc.ABC):
             store.dispatch(
                 NotificationsAddAction(
                     notification=Notification(
-                        title='Speech Recognition',
+                        title=self.name,
                         content=(
-                            f'An error occurred while running the "{self.name}" '
-                            'speech recognition engine. Please check the logs '
-                            'for more details.'
+                            f'An error occurred while running "{self.name}".'
+                            'Please check the logs for more details.'
                         ),
                         color=DANGER_COLOR,
                     ),
@@ -81,25 +70,25 @@ class BackgroundRunningMixin(abc.ABC):
 
     @abc.abstractmethod
     async def _run(self) -> None:
-        """Run the recognition engine."""
         msg = 'This method should be implemented by subclasses.'
         raise NotImplementedError(msg)
 
-    def run(self) -> None:
-        """Run the recognition engine in a background task."""
+    def run(self) -> bool:
+        """Run the engine if it is not already running."""
         with self._run_lock:
             if self._is_running:
-                return
+                return True
             self._is_running = True
             create_task(
                 self._run(),
                 callback=self._set_task,
                 name='VoskEngine.run',
             )
+            return True
 
     @final
     def stop(self) -> None:
-        """Stop the recognition engine."""
+        """Stop the engine if it is running."""
         if not self._task:
             return
         self._task.cancel()
@@ -115,13 +104,3 @@ class BackgroundRunningMixin(abc.ABC):
             self.run()
         else:
             self.stop()
-
-    async def report(self, result: str) -> None:
-        """Report the recognized speech."""
-        logger.debug(
-            'Unprocessed speech recognized',
-            extra={
-                'result': result,
-                'engine_name': self.name,
-            },
-        )