+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - fix(system-manager): better handle `None` output of handlers in response to `send_command`
 - refactor: replace `default_factory` with `default` whenever possible so that `default_provider` can be used
 - refactor(core): improve importing modules in service threads, making it almost identical to python environment, at the cost of patching `importlib._bootstrap._find_and_load_unlocked`
+- refactor(speech-recognition): move the `BackgroundRunningMixin` and `NeedsSetupMixin` abstractions to the core as other modules can use them too
 
 ## Version 1.4.0